+++ src/main/resources/db/migration/V15__Create_AuditLogs_Table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS audit_logs (
-    id INT AUTO_INCREMENT PRIMARY KEY,
-    user_id INT,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    user_id BIGINT,
     action VARCHAR(255),
     description TEXT,
     timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
