+++ src/main/resources/db/migration/V16__Create_FileAttachments_Table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS file_attachments (
-    id INT AUTO_INCREMENT PRIMARY KEY,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
     entity_type VARCHAR(255) COMMENT 'vd: medical_record, invoice',
-    entity_id INT,
+    entity_id BIGINT,
     file_path VARCHAR(255),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
