+++ src/main/resources/db/migration/V19__Create_DoctorSchedule_Table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS doctor_schedule (
-    id INT AUTO_INCREMENT PRIMARY KEY,
-    doctor_id INT,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    doctor_id BIGINT,
     day_of_week VARCHAR(20) COMMENT 'Monday, Tuesday,…',
     start_time TIME,
     end_time TIME,
