+++ src/main/resources/db/migration/V23__Create_PrecriptionTemplate_Table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS prescription_template (
-    id INT AUTO_INCREMENT PRIMARY KEY,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(255),
     description TEXT,
     content TEXT,
