+++ src/main/resources/db/migration/V6__Create_Doctors_Table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS doctors (
-    id INT AUTO_INCREMENT PRIMARY KEY,
-    user_id INT UNIQUE,
-    department_id INT,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    user_id BIGINT UNIQUE,
+    department_id BIGINT,
     specialization VARCHAR(255),
     license_number VARCHAR(255),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
